requirements.txt
altgraph==0.17.2
cx-Freeze==6.8.1
fpdf==1.7.2
idna==3.3
imageio==2.10.1
importlib-metadata==4.8.1
numpy==1.21.3
opencv-python==4.5.4.58
Pillow==8.4.0
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
zipp==3.6.0
